+++ doc/changelog.rst
@@ -50,6 +50,10 @@ General
     <https://github.com/deepmind/mujoco/tree/main/test/benchmark/engine_core_smooth_benchmark_test.cc>`_.
   - 50% faster ``mju_dotSparse`` using manual loop unroll. See `engine_util_sparse_benchmark_test
     <https://github.com/deepmind/mujoco/tree/main/test/benchmark/engine_util_sparse_benchmark_test.cc>`_.
+- Add API functions ``mj_loadPluginLibrary`` and  ``mj_loadAllPluginLibraries``. The first function is identical to
+  ``dlopen`` on a POSIX system, and to ``LoadLibraryA`` on Windows. The second function scans a specified directory for
+  all dynamic libraries file and loads each library found. Dynamic libraries opened by these functions are assumed to
+  register one or more MuJoCo plugins on load.
 
 Simulate
 ^^^^^^^^